+++ src/TemperatureDataBaseWPF/Models/DataBaseModel.cs
@@ -1,0 +1,13 @@
+﻿namespace TemperatureDataBaseWPF.Models
+{
+    public class DataBaseModel
+    {
+        public float Temperature {  get; set; }
+        public DateTime Time { get; set; }
+        public DataBaseModel()
+        {
+            Temperature = 0;
+            Time = DateTime.Now;
+        }
+    }
+}